requirements.txt
Flask==2.2.3

awscli~=1.27.114
boto3~=1.26.114
python-jenkins~=1.8.0
pip~=23.1
wheel~=0.38.4
rsa~=4.7.2
pyasn1~=0.4.8
PyYAML~=5.4.1
setuptools~=65.5.1
botocore~=1.29.114
docutils~=0.16
colorama~=0.4.4
python-dateutil~=2.8.2
s3transfer~=0.6.0
jmespath~=1.0.1
six~=1.16.0
urllib3~=1.26.15
certifi~=2022.12.7
packaging~=23.1
alembic~=1.10.3
SQLAlchemy~=2.0.9